pyproject.toml
[project]
name = "fuel-project"
version = "0.1.0"
description = "Fuel route optimizer"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dj-database-url>=3.1.0",
    "django>=6.0.2",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "geopy>=2.4.1",
    "openrouteservice>=2.3.3",
    "pandas>=3.0.0",
    "pre-commit>=4.5.1",
    "psycopg>=3.3.2",
    "pygeocodio>=2.0.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.1",
    "requests>=2.32.5",
]
